requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-core==2.18.2
starlette==0.37.2
pyyaml>=6.0

# LLM clients
openai>=1.0.0

# HTTP client
httpx==0.27.0
sse-starlette

# Database (optional)
SQLAlchemy==2.0.30
PyMySQL==1.1.1

# Utilities
dateparser==1.2.0
tzdata==2024.1

# Memory (for Phase 2 - will be added later)
# sentence-transformers>=2.2.0
# numpy>=1.24.0
# scikit-learn>=1.3.0